removed free products from topten replaced $this->loadHelper('image') by if class exists (FE completly, some in BE) removed some old loadhelper pointing to non existing helpers added hash based cache for AvaTax
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/branches/com_virtuemart2.0.14mp@6909 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/install/install.sql
@@ -733,7 +733,7 @@ CREATE TABLE IF NOT EXISTS `#__virtuemart_products` (
   `metaauthor` varchar(400),
   `layout` char(16),
   `published` tinyint(1),
-  `pordering` smallint(2) UNSIGNED NOT NULL DEFAULT '0',
+  `pordering` mediumint(2) UNSIGNED NOT NULL DEFAULT '0',
   `created_on` datetime NOT NULL default '0000-00-00 00:00:00',
   `created_by` int(11) NOT NULL DEFAULT '0',
   `modified_on` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
